tab-bar: タブバーで画面の切り替えが一応、できるようになりました。
